separated faulty agent tests
